ui improvement and icon updated
