+++ app/(main)/(drawer)/(tabs)/_layout.tsx
@@ -6,22 +6,16 @@ import { useClientOnlyValue } from "@/components/useClientOnlyValue";
 import { useBreakpoints } from "@/hooks";
 import { useTheme } from "@react-navigation/native";
 import NotificationIcon from "@/components/notifications/notification-icon";
-import { FontAwesome6 } from "@expo/vector-icons";
-
-const TabBarIcon = (props: {
-  name: React.ComponentProps<typeof FontAwesome6>["name"];
-  color: string;
-}) => {
-  return (
-    <FontAwesome6
-      size={28}
-      style={{
-        marginBottom: -3,
-      }}
-      {...props}
-    />
-  );
-};
+import { FontAwesomeIcon } from "@fortawesome/react-native-fontawesome";
+import {
+  faHandshake,
+  faComment,
+  faStar,
+  faUser,
+} from "@fortawesome/free-regular-svg-icons";
+import {
+  faFileSignature,
+} from "@fortawesome/free-solid-svg-icons";
 
 const TabLayout = () => {
   const { xl } = useBreakpoints();
@@ -52,7 +46,11 @@ const TabLayout = () => {
           title: "Proposals",
           tabBarLabel: "Proposals",
           tabBarIcon: ({ color }) => (
-            <TabBarIcon name="handshake-simple" color={color} />
+            <FontAwesomeIcon
+              color={color}
+              icon={faHandshake}
+              size={28}
+            />
           ),
           headerRight: () => <NotificationIcon />,
         }}
@@ -63,7 +61,11 @@ const TabLayout = () => {
           title: "Messages",
           tabBarLabel: "Messages",
           tabBarIcon: ({ color }) => (
-            <TabBarIcon name="comment" color={color} />
+            <FontAwesomeIcon
+              color={color}
+              icon={faComment}
+              size={24}
+            />
           ),
           headerRight: () => <NotificationIcon />,
         }}
@@ -74,14 +76,24 @@ const TabLayout = () => {
           title: "Collaborations",
           tabBarLabel: "Collaborations",
           headerRight: () => <NotificationIcon />,
-          tabBarIcon: ({ color }) => <TabBarIcon name="star" color={color} />,
+          tabBarIcon: ({ color }) => (
+            <FontAwesomeIcon
+              color={color}
+              icon={faStar}
+              size={24}
+            />
+          ),
         }}
       />
       <Tabs.Screen
         name="contracts"
         options={{
           tabBarIcon: ({ color }) => (
-            <TabBarIcon name="file-signature" color={color} />
+            <FontAwesomeIcon
+              color={color}
+              icon={faFileSignature}
+              size={24}
+            />
           ),
           title: "Contracts",
           tabBarLabel: "Contracts",
@@ -95,7 +107,11 @@ const TabLayout = () => {
           title: "Profile",
           tabBarLabel: "Profile",
           tabBarIcon: ({ color }) => (
-            <TabBarIcon name="user" color={color} />
+            <FontAwesomeIcon
+              color={color}
+              icon={faUser}
+              size={24}
+            />
           ),
         }}
       />
