+++ app/(main)/(screens)/(apply-now)/apply-now/[pageID].tsx
@@ -21,7 +21,6 @@ import {
   Appbar,
   Button,
   Card,
-  Chip,
   HelperText,
   IconButton,
   List,
