+++ app/(main)/(screens)/(apply-now)/apply-now/preview.tsx
