--- app/(main)/(screens)/collaboration-details.tsx
@@ -1,9 +1,0 @@
-import CollaborationPage from "@/components/collaboration/CollaborationDetails";
-
-const CollaborationDetailsScreen = () => {
-  return (
-    <CollaborationPage />
-  );
-};
-
-export default CollaborationDetailsScreen;
